+++ cmd/server/config.go
@@ -10,7 +10,8 @@ import (
 )
 
 type llmConfig interface {
-	llm() (handlers.LLM, error)
+	llm(string) (handlers.LLM, error)
+	titleGen(string) (handlers.TitleGenerator, error)
 }
 
 // BaseLLMConfig contains the common fields for all LLM configurations.
@@ -20,10 +21,12 @@ type BaseLLMConfig struct {
 }
 
 type config struct {
-	Port            string                          `yaml:"port"`
-	LLM             llmConfig                       `yaml:"llm"`
-	MCPSSEServers   map[string]mcpSSEServerConfig   `yaml:"mcpSSEServers"`
-	MCPStdIOServers map[string]mcpStdIOServerConfig `yaml:"mcpStdIOServers"`
+	Port                 string                          `yaml:"port"`
+	SystemPrompt         string                          `yaml:"systemPrompt"`
+	TitleGeneratorPrompt string                          `yaml:"titleGeneratorPrompt"`
+	LLM                  llmConfig                       `yaml:"llm"`
+	MCPSSEServers        map[string]mcpSSEServerConfig   `yaml:"mcpSSEServers"`
+	MCPStdIOServers      map[string]mcpStdIOServerConfig `yaml:"mcpStdIOServers"`
 }
 
 type ollamaConfig struct {
@@ -91,29 +94,45 @@ func (c *config) UnmarshalYAML(value *yaml.Node) error {
 	return nil
 }
 
-func (o ollamaConfig) llm() (handlers.LLM, error) {
+func (o ollamaConfig) newOllama(systemPrompt string) (services.Ollama, error) {
 	if o.Model == "" {
-		return nil, fmt.Errorf("model is required")
+		return services.Ollama{}, fmt.Errorf("model is required")
 	}
 
 	host := o.Host
 	if host == "" {
 		host = os.Getenv("OLLAMA_HOST")
 	}
-	return services.NewOllama(host, o.Model), nil
+	return services.NewOllama(host, o.Model, systemPrompt), nil
 }
 
-func (a anthropicConfig) llm() (handlers.LLM, error) {
+func (o ollamaConfig) llm(systemPrompt string) (handlers.LLM, error) {
+	return o.newOllama(systemPrompt)
+}
+
+func (o ollamaConfig) titleGen(systemPrompt string) (handlers.TitleGenerator, error) {
+	return o.newOllama(systemPrompt)
+}
+
+func (a anthropicConfig) newAnthropic(systemPrompt string) (services.Anthropic, error) {
 	if a.Model == "" {
-		return nil, fmt.Errorf("model is required")
+		return services.Anthropic{}, fmt.Errorf("model is required")
 	}
 	if a.MaxTokens == 0 {
-		return nil, fmt.Errorf("max_tokens is required")
+		return services.Anthropic{}, fmt.Errorf("max_tokens is required")
 	}
 
 	apiKey := a.APIKey
 	if apiKey == "" {
 		apiKey = os.Getenv("ANTHROPIC_API_KEY")
 	}
-	return services.NewAnthropic(apiKey, a.Model, a.MaxTokens), nil
+	return services.NewAnthropic(apiKey, a.Model, systemPrompt, a.MaxTokens), nil
+}
+
+func (a anthropicConfig) llm(systemPrompt string) (handlers.LLM, error) {
+	return a.newAnthropic(systemPrompt)
+}
+
+func (a anthropicConfig) titleGen(systemPrompt string) (handlers.TitleGenerator, error) {
+	return a.newAnthropic(systemPrompt)
 }